+++ src/control_plane/migration/mongodb/mod.rs
@@ -2,9 +2,13 @@ mod backup_mongodb_database;
 mod mongodb_backup_options;
 mod mongodb_connection_uri;
 mod mongodb_restore_options;
+mod mongodb_verify_target_options;
 mod restore_mongodb_database;
+mod verify_mongodb_target;
 
 pub(crate) use backup_mongodb_database::backup_mongodb_database;
 pub(crate) use mongodb_backup_options::MongoDbBackupOptions;
 pub(crate) use mongodb_restore_options::MongoDbRestoreOptions;
+pub(crate) use mongodb_verify_target_options::MongoDbVerifyTargetOptions;
 pub(crate) use restore_mongodb_database::restore_mongodb_database;
+pub(crate) use verify_mongodb_target::verify_mongodb_target;